+++ examples/mqtt.rs
@@ -11,7 +11,9 @@ struct NestedSettings {
 
 #[derive(Clone, Default, Miniconf, Debug)]
 struct Settings {
+    #[miniconf(defer)]
     inner: NestedSettings,
+    #[miniconf(defer)]
     amplitude: [f32; 2],
     exit: bool,
 }