+++ py/miniconf-mqtt/miniconf/miniconf.py
@@ -35,12 +35,11 @@ class Miniconf:
             client: A connected MQTT5 client.
             prefix: The MQTT toptic prefix of the device to control.
         """
-        self.request_id = 0
         self.client = client
         self.prefix = prefix
         self.inflight = {}
         self.client.on_message = self._handle_response
-        self.response_topic = f'{prefix}/response/{uuid.uuid1().hex}'
+        self.response_topic = f'{prefix}/response'
         self.client.subscribe(self.response_topic)
 
     def _handle_response(self, _client, topic, payload, _qos, properties):
@@ -55,15 +54,18 @@ class Miniconf:
         """
         if topic == self.response_topic:
             # Extract request_id corrleation data from the properties
-            request_id = int.from_bytes(
-                properties['correlation_data'][0], 'big')
+            request_id = properties['correlation_data'][0]
+
+            if request_id not in self.inflight:
+                LOGGER.info("Discarding message with CD: %s", request_id)
+                return
 
             self.inflight[request_id].set_result(json.loads(payload))
             del self.inflight[request_id]
         else:
             LOGGER.warning('Unexpected message on "%s"', topic)
 
-    async def command(self, path, value, retain=True):
+    async def command(self, path, value, retain=False, timeout=5):
         """Write the provided data to the specified path.
 
         Args:
@@ -71,6 +73,7 @@ class Miniconf:
             value: The value to write to the path.
             retain: Retain the MQTT message changing the setting
                 by the broker.
+            timeout: The maximum time to wait for the response in seconds.
 
         Returns:
             The response to the command as a dictionary.
@@ -80,19 +83,18 @@ class Miniconf:
         fut = asyncio.get_running_loop().create_future()
 
         # Assign unique correlation data for response dispatch
-        assert self.request_id not in self.inflight
-        self.inflight[self.request_id] = fut
-        correlation_data = self.request_id.to_bytes(4, 'big')
-        self.request_id += 1
+        request_id = uuid.uuid1().hex.encode()
+        assert request_id not in self.inflight
+        self.inflight[request_id] = fut
 
         payload = json.dumps(value, separators=(",", ":"))
-        LOGGER.info('Sending "%s" to "%s"', value, topic)
+        LOGGER.info('Sending "%s" to "%s" with CD: %s', value, topic, request_id)
 
         self.client.publish(
             topic, payload=payload, qos=0, retain=retain,
             response_topic=self.response_topic,
-            correlation_data=correlation_data)
+            correlation_data=request_id)
 
-        result = await fut
+        result = await asyncio.wait_for(fut, timeout)
         if result['code'] != 0:
             raise MiniconfException(result['msg'])