+++ src/iter.rs
@@ -1,23 +1,31 @@
 use super::Miniconf;
+use core::marker::PhantomData;
 use heapless::String;
 
-pub struct MiniconfIter<'a, Settings: Miniconf + ?Sized, const TS: usize> {
-    pub(crate) settings: &'a Settings,
-    pub(crate) state: &'a mut [usize],
+/// An iterator over the paths in a Miniconf namespace.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub struct MiniconfIter<M: ?Sized, const L: usize, const TS: usize> {
+    marker: PhantomData<M>,
+    state: [usize; L],
 }
 
-impl<'a, Settings: Miniconf + ?Sized, const TS: usize> Iterator for MiniconfIter<'a, Settings, TS> {
+impl<M: ?Sized, const L: usize, const TS: usize> Default for MiniconfIter<M, L, TS> {
+    fn default() -> Self {
+        MiniconfIter {
+            marker: PhantomData,
+            state: [0; L],
+        }
+    }
+}
+
+impl<M: Miniconf + ?Sized, const L: usize, const TS: usize> Iterator for MiniconfIter<M, L, TS> {
     type Item = String<TS>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        let mut topic_buffer: String<TS> = String::new();
+        let mut path = Self::Item::new();
 
-        if self
-            .settings
-            .recurse_paths(self.state, &mut topic_buffer)
-            .is_some()
-        {
-            Some(topic_buffer)
+        if M::next_path(&mut self.state, &mut path) {
+            Some(path)
         } else {
             None
         }