+++ tests/generics.rs
@@ -1,43 +1,40 @@
-use miniconf::{Miniconf, MiniconfAtomic};
+use miniconf::Miniconf;
 use serde::{Deserialize, Serialize};
 
 #[test]
 fn generic_type() {
     #[derive(Miniconf, Default)]
-    struct Settings<T: Miniconf> {
+    struct Settings<T> {
         pub data: T,
     }
 
     let mut settings = Settings::<f32>::default();
-    settings
-        .string_set("data".split('/').peekable(), b"3.0")
-        .unwrap();
+    settings.set("data", b"3.0").unwrap();
     assert_eq!(settings.data, 3.0);
 
     // Test metadata
-    let metadata = settings.get_metadata();
-    assert_eq!(metadata.max_depth, 2);
-    assert_eq!(metadata.max_topic_size, "data".len());
+    let metadata = Settings::<f32>::metadata();
+    assert_eq!(metadata.max_depth, 1);
+    assert_eq!(metadata.max_length, "data".len());
 }
 
 #[test]
 fn generic_array() {
     #[derive(Miniconf, Default)]
-    struct Settings<T: Miniconf> {
+    struct Settings<T> {
+        #[miniconf(defer)]
         pub data: [T; 2],
     }
 
     let mut settings = Settings::<f32>::default();
-    settings
-        .string_set("data/0".split('/').peekable(), b"3.0")
-        .unwrap();
+    settings.set("data/0", b"3.0").unwrap();
 
     assert_eq!(settings.data[0], 3.0);
 
     // Test metadata
-    let metadata = settings.get_metadata();
-    assert_eq!(metadata.max_depth, 3);
-    assert_eq!(metadata.max_topic_size, "data/0".len());
+    let metadata = Settings::<f32>::metadata();
+    assert_eq!(metadata.max_depth, 2);
+    assert_eq!(metadata.max_length, "data/0".len());
 }
 
 #[test]
@@ -47,22 +44,20 @@ fn generic_struct() {
         pub inner: T,
     }
 
-    #[derive(MiniconfAtomic, Deserialize, Serialize, Default)]
+    #[derive(Serialize, Deserialize, Default)]
     struct Inner {
         pub data: f32,
     }
 
     let mut settings = Settings::<Inner>::default();
-    settings
-        .string_set("inner".split('/').peekable(), b"{\"data\": 3.0}")
-        .unwrap();
+    settings.set("inner", b"{\"data\": 3.0}").unwrap();
 
     assert_eq!(settings.inner.data, 3.0);
 
     // Test metadata
-    let metadata = settings.get_metadata();
-    assert_eq!(metadata.max_depth, 2);
-    assert_eq!(metadata.max_topic_size, "inner".len());
+    let metadata = Settings::<Inner>::metadata();
+    assert_eq!(metadata.max_depth, 1);
+    assert_eq!(metadata.max_length, "inner".len());
 }
 
 #[test]
@@ -72,23 +67,20 @@ fn generic_atomic() {
         pub atomic: Inner<T>,
     }
 
-    #[derive(Deserialize, Serialize, MiniconfAtomic, Default)]
+    #[derive(Deserialize, Serialize, Default)]
     struct Inner<T> {
         pub inner: [T; 5],
     }
 
     let mut settings = Settings::<f32>::default();
     settings
-        .string_set(
-            "atomic".split('/').peekable(),
-            b"{\"inner\": [3.0, 0, 0, 0, 0]}",
-        )
+        .set("atomic", b"{\"inner\": [3.0, 0, 0, 0, 0]}")
         .unwrap();
 
     assert_eq!(settings.atomic.inner[0], 3.0);
 
     // Test metadata
-    let metadata = settings.get_metadata();
-    assert_eq!(metadata.max_depth, 2);
-    assert_eq!(metadata.max_topic_size, "atomic".len());
+    let metadata = Settings::<f32>::metadata();
+    assert_eq!(metadata.max_depth, 1);
+    assert_eq!(metadata.max_length, "atomic".len());
 }